Changed Schema and Relocated timeSlot functions
